+++ src/app/selectors.ts
@@ -1,5 +1,4 @@
 import { createSelector } from '@reduxjs/toolkit';
-import { DEAFULT_CHARACTER } from 'features/character/characterListSlice';
 import { RootState } from 'app/rootReducer';
 
 const getCharacterList = (state: RootState) => state.characterList;
@@ -8,9 +7,6 @@ const getSelectedCharacterId = (state: RootState) => state.selectedCharacter;
 export const getSelectedCharacter = createSelector(
   [getCharacterList, getSelectedCharacterId],
   (characterList, selectedCharacterId) => {
-    return (
-      characterList.list.find(char => char.id === selectedCharacterId) ||
-      DEAFULT_CHARACTER
-    );
+    return characterList.list.find(char => char.id === selectedCharacterId);
   },
 );
