+++ src/features/character/characterListSlice.ts
@@ -142,268 +142,269 @@ export type CharacterListItem = CharacterBase &
 // }
 export type CharacterList = {
   id: string;
-  loading: 'idle' | 'pending';
+  loading: 'idle' | 'pending' | 'error';
   list: Array<CharacterListItem>;
+  error?: string;
 };
-export const DEFAULT_ID = 'id-DEFAULT';
-export const DEAFULT_CHARACTER: CharacterListItem = {
-  id: DEFAULT_ID,
-  allSources: false,
-  raceData: {
-    race: 'Human',
-    chosenRaceAbilities: [],
-    standardRaceAbilities: [
-      {
-        str: 1,
-        dex: 1,
-        con: 1,
-        int: 1,
-        wis: 1,
-        cha: 1,
-      },
-    ],
-    chosenRaceSkillProficiencies: [],
-    standardRaceSkillProficiencies: [],
-    chosenRaceLanguages: ['celestial'],
-    standardRaceLanguages: ['common'],
-  },
-  classData: {
-    classElement: 'Fighter',
-    subClass: 'Champion',
-    chosenClassSkillProficiencies: ['acrobatics', 'athletics'],
-    standardClassArmorProficiencies: ['light', 'medium', 'heavy', 'shields'],
-    standardClassWeaponProficiencies: ['simple', 'martial'],
-    standardClassToolProficiencies: [],
-    abilityScores: {
-      rollMethod: 'standard',
-      str: 15,
-      dex: 13,
-      con: 14,
-      int: 10,
-      wis: 8,
-      cha: 12,
-    },
-  },
-  descriptionData: {
-    name: 'Regdar Ridgeheart',
-    background: 'Acolyte',
-    alignment: 'LG',
-    characteristicsSource: 'Acolyte',
-    imageUrl: '/img/races/human.png',
-    hair: 'Blonde',
-    skin: 'Fair',
-    eyes: 'Brown',
-    height: '193',
-    weight: '89',
-    age: '34',
-    backstory:
-      'You squired for a knight who taught you how to fight, care for a steed, and conduct yourself with honor. You decided to take up that path for yourself.\n\nMan-at-arms: You were the member of an elite cavalry unit of an army, trained in heavy armor and a variety of weapons to charge the enemy and crush the opposition.',
-    chosenBackgroundSkillProficiencies: [],
-    standardBackgroundSkillProficiencies: ['insight', 'religion'],
-    chosenBackgroundToolProficiencies: [],
-    standardBackgroundToolProficiencies: [],
-    chosenBackgroundLanguages: ['dwarvish', 'halfling'],
-    standardBackgroundLanguages: [],
-    characteristicsPersonalityTrait:
-      "I idolize a particular hero of my faith, and constantly refer to that person's deeds and example.",
-    characteristicsIdeal:
-      'Charity. I always try to help those in need, no matter what the personal cost. (Good)',
-    characteristicsBond: 'Everything I do is for the common people.',
-    characteristicsFlaw:
-      'Once I pick a goal, I become obsessed with it to the detriment of everything else in my life.',
-  },
-  equipmentData: {
-    items: ['Longsword', 'Chain Mail', 'Shield', 'Arrows (20)'],
-  },
-  customData: {
-    customAbilities: [],
-    customSkillProficiencies: [],
-    customArmorProficiencies: [],
-    customWeaponProficiencies: [],
-    customToolProficiencies: [],
-    customLanguages: [],
-  },
-  gameData: {
-    level: 1,
-    feats: [],
-    conditions: ['invisible'],
-    defenses: [{ type: DefenseType.Immunity, name: 'cold' }],
-    spells: [],
-    attunements: [],
-    actions: [],
-    extras: [],
-    ac: 0,
-    currentHp: 12,
-    currentHd: 1,
-  },
-};
+// export const DEFAULT_ID = 'id-DEFAULT';
+// export const DEAFULT_CHARACTER: CharacterListItem = {
+//   id: DEFAULT_ID,
+//   allSources: false,
+//   raceData: {
+//     race: 'Human',
+//     chosenRaceAbilities: [],
+//     standardRaceAbilities: [
+//       {
+//         str: 1,
+//         dex: 1,
+//         con: 1,
+//         int: 1,
+//         wis: 1,
+//         cha: 1,
+//       },
+//     ],
+//     chosenRaceSkillProficiencies: [],
+//     standardRaceSkillProficiencies: [],
+//     chosenRaceLanguages: ['celestial'],
+//     standardRaceLanguages: ['common'],
+//   },
+//   classData: {
+//     classElement: 'Fighter',
+//     subClass: 'Champion',
+//     chosenClassSkillProficiencies: ['acrobatics', 'athletics'],
+//     standardClassArmorProficiencies: ['light', 'medium', 'heavy', 'shields'],
+//     standardClassWeaponProficiencies: ['simple', 'martial'],
+//     standardClassToolProficiencies: [],
+//     abilityScores: {
+//       rollMethod: 'standard',
+//       str: 15,
+//       dex: 13,
+//       con: 14,
+//       int: 10,
+//       wis: 8,
+//       cha: 12,
+//     },
+//   },
+//   descriptionData: {
+//     name: 'Regdar Ridgeheart',
+//     background: 'Acolyte',
+//     alignment: 'LG',
+//     characteristicsSource: 'Acolyte',
+//     imageUrl: '/img/races/human.png',
+//     hair: 'Blonde',
+//     skin: 'Fair',
+//     eyes: 'Brown',
+//     height: '193',
+//     weight: '89',
+//     age: '34',
+//     backstory:
+//       'You squired for a knight who taught you how to fight, care for a steed, and conduct yourself with honor. You decided to take up that path for yourself.\n\nMan-at-arms: You were the member of an elite cavalry unit of an army, trained in heavy armor and a variety of weapons to charge the enemy and crush the opposition.',
+//     chosenBackgroundSkillProficiencies: [],
+//     standardBackgroundSkillProficiencies: ['insight', 'religion'],
+//     chosenBackgroundToolProficiencies: [],
+//     standardBackgroundToolProficiencies: [],
+//     chosenBackgroundLanguages: ['dwarvish', 'halfling'],
+//     standardBackgroundLanguages: [],
+//     characteristicsPersonalityTrait:
+//       "I idolize a particular hero of my faith, and constantly refer to that person's deeds and example.",
+//     characteristicsIdeal:
+//       'Charity. I always try to help those in need, no matter what the personal cost. (Good)',
+//     characteristicsBond: 'Everything I do is for the common people.',
+//     characteristicsFlaw:
+//       'Once I pick a goal, I become obsessed with it to the detriment of everything else in my life.',
+//   },
+//   equipmentData: {
+//     items: ['Longsword', 'Chain Mail', 'Shield', 'Arrows (20)'],
+//   },
+//   customData: {
+//     customAbilities: [],
+//     customSkillProficiencies: [],
+//     customArmorProficiencies: [],
+//     customWeaponProficiencies: [],
+//     customToolProficiencies: [],
+//     customLanguages: [],
+//   },
+//   gameData: {
+//     level: 1,
+//     feats: [],
+//     conditions: ['invisible'],
+//     defenses: [{ type: DefenseType.Immunity, name: 'cold' }],
+//     spells: [],
+//     attunements: [],
+//     actions: [],
+//     extras: [],
+//     ac: 0,
+//     currentHp: 12,
+//     currentHd: 1,
+//   },
+// };
 
-const MOE: CharacterListItem = {
-  id: 'id-cb5de106a9e47',
-  allSources: true,
-  raceData: {
-    race: 'Halfling (Ghostwise)',
-    chosenRaceAbilities: [],
-    standardRaceAbilities: [
-      {
-        dex: 2,
-        wis: 1,
-      },
-    ],
-    chosenRaceSkillProficiencies: [],
-    standardRaceSkillProficiencies: [],
-    chosenRaceLanguages: [],
-    standardRaceLanguages: ['common', 'halfling'],
-  },
-  classData: {
-    classElement: 'Druid',
-    subClass: 'Circle of the Shepherd',
-    chosenClassSkillProficiencies: ['insight', 'perception'],
-    standardClassArmorProficiencies: ['light', 'medium', 'shields'],
-    standardClassWeaponProficiencies: [
-      'clubs',
-      'daggers',
-      'darts',
-      'javelins',
-      'maces',
-      'quarterstaffs',
-      'scimitars',
-      'sickles',
-      'slings',
-      'spears',
-    ],
-    standardClassToolProficiencies: [],
-    abilityScores: {
-      rollMethod: 'rolled',
-      str: 8,
-      dex: 15,
-      con: 17,
-      int: 11,
-      wis: 17,
-      cha: 11,
-    },
-  },
-  descriptionData: {
-    name: 'Moe Glee, The Minionmancer',
-    background: 'Far Traveler',
-    alignment: 'CG',
-    characteristicsSource: 'Far Traveler',
-    imageUrl: 'https://i.imgur.com/fX2c9M5.png',
-    hair: 'Orange',
-    skin: 'Greenish',
-    eyes: 'Orange',
-    height: '3ft',
-    weight: '40',
-    age: '19',
-    backstory:
-      'Born to a poor farmer family, Moe remembers very little about his parents. He still sees the flames that took them from him in his dreams, violently raging as they consume the last parts of his childhood home.\nMoe can’t remember how the fire started, but it haunts his dreams - night after night he is tormented by these visions. Again and again the flames erupt out of nowhere and each night he fails to escape them.\nShira, the fey spirit that watched over Moe and his tribe often walks the earth in the form of black panther. She managed to pull Moe from the flames on that fateful night, carrying the small boy to safety. She tried going back in for the others but was blocked by the flames.\nOrphaned, and with no-one of his kin left to care for him, Shira decided to bring the boy with her, deep into the forest where he would live with a new kind of tribe.\nFor fifteen years Moe lived together with the animals as a part of their herd. When Moe was seven Shira brought one of her old friends to the camp, a 250 year old tortle druid named Grok. Together Grok and Shira set out to teach Moe about the power of nature, the Great Balance and about the elements that guide it: the power of tooth and claw, of sun and moon and of fire and storm.\nWhen Moe turned 17, Shira came to Moe with a request. She had sensed a growing power, a darkness somewhere far away, disrupting the Great Balance. A force that seeks to hold sway over nature and life. She sent him on a mission to find out what this force is, and to try and restore the balance to their world.\nSo Moe set off towards the foreign lands of Asteria.',
-    chosenBackgroundSkillProficiencies: [],
-    standardBackgroundSkillProficiencies: ['arcana', 'nature'],
-    chosenBackgroundToolProficiencies: ['gaming set'],
-    standardBackgroundToolProficiencies: [],
-    chosenBackgroundLanguages: ['primordial'],
-    standardBackgroundLanguages: [],
-    characteristicsPersonalityTrait:
-      "I have different assumptions from those around me concerning personal space, blithely invading others' space in innocence, or reacting to ignorant invasion of my own.",
-    characteristicsIdeal:
-      "Adventure. I'm far from home, and everything is strange and wonderful! (Chaotic)",
-    characteristicsBond:
-      'Though I had no choice, I lament having to leave my loved one(s) behind. I hope to see them again one day.',
-    characteristicsFlaw:
-      'I have a weakness for the new intoxicants and other pleasures of this land.',
-  },
-  equipmentData: {
-    items: [
-      'Quarterstaff',
-      'Scimitar',
-      'Leather Armor',
-      'Shield',
-      "Explorer's Pack",
-      'Totem',
-    ],
-  },
-  customData: {
-    customAbilities: [{ wis: 2 }],
-    customSkillProficiencies: [],
-    customArmorProficiencies: [],
-    customWeaponProficiencies: [],
-    customToolProficiencies: [],
-    customLanguages: [],
-  },
-  gameData: {
-    level: 4,
-    feats: ['Resilient'],
-    spells: [
-      {
-        row: 0,
-        name: 'Acid Splash',
-      },
-      {
-        row: 1,
-        name: 'Animal Friendship',
-      },
-      {
-        row: 12,
-        name: 'Call Lightning',
-      },
-      {
-        row: 17,
-        name: 'Conjure Animals',
-      },
-      {
-        row: 40,
-        name: 'Entangle',
-      },
-      {
-        row: 49,
-        name: 'Flaming Sphere',
-      },
-      {
-        row: 56,
-        name: 'Goodberry',
-      },
-      {
-        row: 59,
-        name: 'Guidance',
-      },
-      {
-        row: 80,
-        name: 'Healing Spirit',
-      },
-      {
-        row: 101,
-        name: 'Moonbeam',
-      },
-      {
-        row: 105,
-        name: 'Shillelagh',
-      },
-      {
-        row: 144,
-        name: 'Spike Growth',
-      },
-    ],
-    conditions: ['invisible'],
-    defenses: [{ type: DefenseType.Immunity, name: 'cold' }],
-    attunements: [],
-    actions: [],
-    extras: [],
-    ac: 16,
-    currentHp: 12,
-    currentHd: 5,
-    spellSlots: {
-      1: { used: 0 },
-      2: { used: 0 },
-      3: { used: 0 },
-      4: { used: 0 },
-      5: { used: 0 },
-      6: { used: 0 },
-      7: { used: 0 },
-      8: { used: 0 },
-      9: { used: 0 },
-    },
-  },
-};
+// const MOE: CharacterListItem = {
+//   id: 'id-cb5de106a9e47',
+//   allSources: true,
+//   raceData: {
+//     race: 'Halfling (Ghostwise)',
+//     chosenRaceAbilities: [],
+//     standardRaceAbilities: [
+//       {
+//         dex: 2,
+//         wis: 1,
+//       },
+//     ],
+//     chosenRaceSkillProficiencies: [],
+//     standardRaceSkillProficiencies: [],
+//     chosenRaceLanguages: [],
+//     standardRaceLanguages: ['common', 'halfling'],
+//   },
+//   classData: {
+//     classElement: 'Druid',
+//     subClass: 'Circle of the Shepherd',
+//     chosenClassSkillProficiencies: ['insight', 'perception'],
+//     standardClassArmorProficiencies: ['light', 'medium', 'shields'],
+//     standardClassWeaponProficiencies: [
+//       'clubs',
+//       'daggers',
+//       'darts',
+//       'javelins',
+//       'maces',
+//       'quarterstaffs',
+//       'scimitars',
+//       'sickles',
+//       'slings',
+//       'spears',
+//     ],
+//     standardClassToolProficiencies: [],
+//     abilityScores: {
+//       rollMethod: 'rolled',
+//       str: 8,
+//       dex: 15,
+//       con: 17,
+//       int: 11,
+//       wis: 17,
+//       cha: 11,
+//     },
+//   },
+//   descriptionData: {
+//     name: 'Moe Glee, The Minionmancer',
+//     background: 'Far Traveler',
+//     alignment: 'CG',
+//     characteristicsSource: 'Far Traveler',
+//     imageUrl: 'https://i.imgur.com/fX2c9M5.png',
+//     hair: 'Orange',
+//     skin: 'Greenish',
+//     eyes: 'Orange',
+//     height: '3ft',
+//     weight: '40',
+//     age: '19',
+//     backstory:
+//       'Born to a poor farmer family, Moe remembers very little about his parents. He still sees the flames that took them from him in his dreams, violently raging as they consume the last parts of his childhood home.\nMoe can’t remember how the fire started, but it haunts his dreams - night after night he is tormented by these visions. Again and again the flames erupt out of nowhere and each night he fails to escape them.\nShira, the fey spirit that watched over Moe and his tribe often walks the earth in the form of black panther. She managed to pull Moe from the flames on that fateful night, carrying the small boy to safety. She tried going back in for the others but was blocked by the flames.\nOrphaned, and with no-one of his kin left to care for him, Shira decided to bring the boy with her, deep into the forest where he would live with a new kind of tribe.\nFor fifteen years Moe lived together with the animals as a part of their herd. When Moe was seven Shira brought one of her old friends to the camp, a 250 year old tortle druid named Grok. Together Grok and Shira set out to teach Moe about the power of nature, the Great Balance and about the elements that guide it: the power of tooth and claw, of sun and moon and of fire and storm.\nWhen Moe turned 17, Shira came to Moe with a request. She had sensed a growing power, a darkness somewhere far away, disrupting the Great Balance. A force that seeks to hold sway over nature and life. She sent him on a mission to find out what this force is, and to try and restore the balance to their world.\nSo Moe set off towards the foreign lands of Asteria.',
+//     chosenBackgroundSkillProficiencies: [],
+//     standardBackgroundSkillProficiencies: ['arcana', 'nature'],
+//     chosenBackgroundToolProficiencies: ['gaming set'],
+//     standardBackgroundToolProficiencies: [],
+//     chosenBackgroundLanguages: ['primordial'],
+//     standardBackgroundLanguages: [],
+//     characteristicsPersonalityTrait:
+//       "I have different assumptions from those around me concerning personal space, blithely invading others' space in innocence, or reacting to ignorant invasion of my own.",
+//     characteristicsIdeal:
+//       "Adventure. I'm far from home, and everything is strange and wonderful! (Chaotic)",
+//     characteristicsBond:
+//       'Though I had no choice, I lament having to leave my loved one(s) behind. I hope to see them again one day.',
+//     characteristicsFlaw:
+//       'I have a weakness for the new intoxicants and other pleasures of this land.',
+//   },
+//   equipmentData: {
+//     items: [
+//       'Quarterstaff',
+//       'Scimitar',
+//       'Leather Armor',
+//       'Shield',
+//       "Explorer's Pack",
+//       'Totem',
+//     ],
+//   },
+//   customData: {
+//     customAbilities: [{ wis: 2 }],
+//     customSkillProficiencies: [],
+//     customArmorProficiencies: [],
+//     customWeaponProficiencies: [],
+//     customToolProficiencies: [],
+//     customLanguages: [],
+//   },
+//   gameData: {
+//     level: 4,
+//     feats: ['Resilient'],
+//     spells: [
+//       {
+//         row: 0,
+//         name: 'Acid Splash',
+//       },
+//       {
+//         row: 1,
+//         name: 'Animal Friendship',
+//       },
+//       {
+//         row: 12,
+//         name: 'Call Lightning',
+//       },
+//       {
+//         row: 17,
+//         name: 'Conjure Animals',
+//       },
+//       {
+//         row: 40,
+//         name: 'Entangle',
+//       },
+//       {
+//         row: 49,
+//         name: 'Flaming Sphere',
+//       },
+//       {
+//         row: 56,
+//         name: 'Goodberry',
+//       },
+//       {
+//         row: 59,
+//         name: 'Guidance',
+//       },
+//       {
+//         row: 80,
+//         name: 'Healing Spirit',
+//       },
+//       {
+//         row: 101,
+//         name: 'Moonbeam',
+//       },
+//       {
+//         row: 105,
+//         name: 'Shillelagh',
+//       },
+//       {
+//         row: 144,
+//         name: 'Spike Growth',
+//       },
+//     ],
+//     conditions: ['invisible'],
+//     defenses: [{ type: DefenseType.Immunity, name: 'cold' }],
+//     attunements: [],
+//     actions: [],
+//     extras: [],
+//     ac: 16,
+//     currentHp: 12,
+//     currentHd: 5,
+//     spellSlots: {
+//       1: { used: 0 },
+//       2: { used: 0 },
+//       3: { used: 0 },
+//       4: { used: 0 },
+//       5: { used: 0 },
+//       6: { used: 0 },
+//       7: { used: 0 },
+//       8: { used: 0 },
+//       9: { used: 0 },
+//     },
+//   },
+// };
 
 export const randomize = () => {
   const id = generateID();
@@ -565,7 +566,7 @@ export const generateID = () =>
 const initialState: CharacterList = {
   id: 'NOT_SAVED_YET',
   loading: 'idle',
-  list: [DEAFULT_CHARACTER, MOE],
+  list: [],
 };
 
 // Async Thunks
@@ -576,27 +577,12 @@ export const backgroundCreate: any = createAsyncThunk<
     dispatch: AppDispatch;
     state: RootState;
   }
->(
-  'characterList/backgroundCreate',
-  async (_, thunkAPI) => {
-    const response = await DnDAboveAPI.create(
-      thunkAPI.getState().characterList,
-    );
-    const data = await response.json();
-    console.log('backgroundCreate', data);
-    return data;
-  },
-  {
-    condition: (_, { getState }) => {
-      const { characterList } = getState();
-      const fetchStatus = characterList.loading;
-      if (fetchStatus === 'pending') {
-        // Already fetching, don't need to re-fetch
-        return false;
-      }
-    },
-  },
-);
+>('characterList/backgroundCreate', async (_, thunkAPI) => {
+  const response = await DnDAboveAPI.create(thunkAPI.getState().characterList);
+  const data = await response.json();
+  console.log('backgroundCreate', data);
+  return data;
+});
 
 export const backgroundSave: any = createAsyncThunk<
   any,
@@ -611,8 +597,9 @@ export const backgroundSave: any = createAsyncThunk<
     const response = await DnDAboveAPI.update(
       thunkAPI.getState().characterList,
     );
-    console.log(response);
-    return await response.json();
+    const data = await response.json();
+    console.log('backgroundSave', data);
+    return data;
   },
   {
     condition: (_, { getState }) => {
@@ -627,7 +614,7 @@ export const backgroundSave: any = createAsyncThunk<
 );
 
 export const getCharacterList = createAsyncThunk<
-  CharacterList,
+  any,
   undefined,
   {
     dispatch: AppDispatch;
@@ -641,10 +628,6 @@ export const getCharacterList = createAsyncThunk<
 
     if (response.status === 404) {
       thunkAPI.dispatch(backgroundCreate());
-    }
-
-    if (response.status === 400) {
-      // Return the known error for future handling
       return thunkAPI.rejectWithValue(await response.json());
     }
     return await response.json();
@@ -777,7 +760,7 @@ const characterListSlice = createSlice({
       }
     },
     removeCharacter(state, action: PayloadAction<string>) {
-      state.list.filter(item => item.id !== action.payload);
+      state.list = state.list.filter(item => item.id !== action.payload);
     },
     setAc(state, action: PayloadAction<{ id: string; ac: number }>) {
       const character = state.list.find(
@@ -1052,14 +1035,15 @@ const characterListSlice = createSlice({
     });
     builder.addCase(backgroundCreate.fulfilled, (state, { payload }) => {
       const { ref } = payload;
-      if (state.loading === 'pending') {
+      if (state.loading === 'pending' || state.loading === 'error') {
         state.loading = 'idle';
         state.id = ref['@ref'].id;
       }
     });
     builder.addCase(backgroundCreate.rejected, (state, { payload }) => {
-      if (state.loading === 'pending') {
-        state.loading = 'idle';
+      if (state.loading === 'pending' || state.loading === 'error') {
+        state.loading = 'error';
+        state.error = (payload as any).message;
       }
     });
     // backgroundSave
@@ -1069,13 +1053,35 @@ const characterListSlice = createSlice({
       }
     });
     builder.addCase(backgroundSave.fulfilled, (state, { payload }) => {
-      if (state.loading === 'pending') {
+      if (state.loading === 'pending' || state.loading === 'error') {
         state.loading = 'idle';
       }
     });
     builder.addCase(backgroundSave.rejected, (state, { payload }) => {
-      if (state.loading === 'pending') {
+      if (state.loading === 'pending' || state.loading === 'error') {
+        state.loading = 'error';
+        state.error = (payload as any).message;
+      }
+    });
+    // getCharacterList
+    builder.addCase(getCharacterList.pending, (state, { payload }) => {
+      if (state.loading === 'idle') {
+        state.loading = 'pending';
+      }
+    });
+    builder.addCase(getCharacterList.fulfilled, (state, { payload }) => {
+      if (state.loading === 'pending' || state.loading === 'error') {
         state.loading = 'idle';
+      }
+      const { ref, data } = payload;
+      console.log('got it', data);
+      state.id = ref['@ref'].id;
+      state.list = data.list;
+    });
+    builder.addCase(getCharacterList.rejected, (state, { payload }) => {
+      if (state.loading === 'pending' || state.loading === 'error') {
+        state.loading = 'error';
+        state.error = (payload as any).message;
       }
     });
   },
