+++ src/features/character/selectedCharacterSlice.ts
@@ -1,9 +1,8 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { DEFAULT_ID } from 'features/character/characterListSlice';
 
 const selectedCharacterSlice = createSlice({
   name: 'character',
-  initialState: DEFAULT_ID,
+  initialState: '',
   reducers: {
     setSelectedCharacter(state, action: PayloadAction<string>) {
       return action.payload;
